+++ docs/COMPLETION_AUDIT.md
@@ -298,6 +298,11 @@ and do not assume Phase 4 is ready without evidence.
   `tests/integration/cli_build/graph_validation_test_command_host_effects/file_reads.rs`,
   leaving env-effect ordering and unrelated executable skip coverage in the
   parent module.
+- Legacy `build-graph build.zen` host-effect integration tests now split
+  declared file-read fallback accept/reject coverage into
+  `tests/integration/cli_build/legacy_graph_command_host_effects/file_reads.rs`,
+  leaving env-effect ordering and unrelated test skip coverage in the parent
+  module.
 - Resolver-backed behavior impl metadata now builds restored impl-block tasks
   once and reuses them for both impl method signature restoration and omitted
   default-method synthesis, preserving the signature-before-defaults ordering.
