+++ tests/integration/cli_build/legacy_graph_command_host_effects.rs
@@ -1,5 +1,8 @@
 use std::process::Command;
 
+#[path = "legacy_graph_command_host_effects/file_reads.rs"]
+mod file_reads;
+
 #[test]
 fn build_graph_command_rejects_undeclared_host_effects() {
     let tmp = tempfile::tempdir().expect("create temp dir");
@@ -194,127 +197,6 @@ main = () i32 {
     assert!(
         !stderr.contains("missing_test.zen"),
         "host-effect validation should run before unrelated test source handling, stderr={stderr}"
-    );
-    assert!(
-        !tmp.path().join("build").exists(),
-        "build-graph command should not start after graph validation fails"
-    );
-}
-
-#[test]
-fn build_graph_command_accepts_declared_file_read_effects() {
-    assert_build_graph_command_accepts_declared_file_read_effect(
-        r#"| .Err { "default" }"#,
-        "build_graph_command_accepts_declared_file_read_effects",
-    );
-}
-
-#[test]
-fn build_graph_command_accepts_wildcard_fallback_declared_file_read_effects() {
-    assert_build_graph_command_accepts_declared_file_read_effect(
-        r#"| _ { "default" }"#,
-        "build_graph_command_accepts_wildcard_fallback_declared_file_read_effects",
-    );
-}
-
-#[test]
-fn build_graph_command_accepts_identifier_fallback_declared_file_read_effects() {
-    assert_build_graph_command_accepts_declared_file_read_effect(
-        r#"| err { "default" }"#,
-        "build_graph_command_accepts_identifier_fallback_declared_file_read_effects",
-    );
-}
-
-fn assert_build_graph_command_accepts_declared_file_read_effect(
-    fallback_arm: &str,
-    case_name: &str,
-) {
-    let tmp = tempfile::tempdir().expect("create temp dir");
-    std::fs::write(
-        tmp.path().join("build.zen"),
-        format!(
-            r#"
-build = (b: Builder) Result<BuildConfig, BuildError> {{
-    manifest = b.os.read_file("build.targets") ?
-        | .Ok(contents) {{ contents }}
-        {fallback_arm}
-    b.add(Executable {{ name: "myapp", main: "main.zen", out_dir: "build/" }})
-    .Ok(b.config())
-}}
-"#,
-        ),
-    )
-    .expect("write build.zen");
-    std::fs::write(tmp.path().join("build.targets"), "myapp\n").expect("write manifest");
-    std::fs::write(
-        tmp.path().join("main.zen"),
-        r#"
-main = () i32 {
-    0
-}
-"#,
-    )
-    .expect("write main.zen");
-
-    let output = Command::new(env!("CARGO_BIN_EXE_zen"))
-        .args(["build-graph", "build.zen"])
-        .current_dir(tmp.path())
-        .output()
-        .expect("run zen build-graph");
-
-    assert!(
-        output.status.success(),
-        "{case_name}: zen build-graph failed: stdout={}, stderr={}",
-        String::from_utf8_lossy(&output.stdout),
-        String::from_utf8_lossy(&output.stderr)
-    );
-
-    let bin_path = tmp.path().join("build").join("myapp");
-    assert!(
-        bin_path.exists(),
-        "expected {} to exist",
-        bin_path.display()
-    );
-    let run = Command::new(&bin_path).output().expect("run built binary");
-    assert!(
-        run.status.success(),
-        "built binary exited with {}",
-        run.status
-    );
-}
-
-#[test]
-fn build_graph_command_rejects_undeclared_file_read_effects_before_execution() {
-    let tmp = tempfile::tempdir().expect("create temp dir");
-    std::fs::write(
-        tmp.path().join("build.zen"),
-        r#"
-build = (b: Builder) Result<BuildConfig, BuildError> {
-    manifest = b.os.read_file("build.targets")
-    b.add(Executable { name: "myapp", main: "main.zen", out_dir: "build/" })
-    .Ok(b.config())
-}
-"#,
-    )
-    .expect("write build.zen");
-
-    let output = Command::new(env!("CARGO_BIN_EXE_zen"))
-        .args(["build-graph", "build.zen"])
-        .current_dir(tmp.path())
-        .output()
-        .expect("run zen build-graph");
-
-    assert!(
-        !output.status.success(),
-        "zen build-graph unexpectedly succeeded: stdout={}, stderr={}",
-        String::from_utf8_lossy(&output.stdout),
-        String::from_utf8_lossy(&output.stderr)
-    );
-    assert!(
-        String::from_utf8_lossy(&output.stderr)
-            .contains("undeclared host effect: read file `build.targets`"),
-        "expected undeclared file read diagnostic, stderr={}",
-        String::from_utf8_lossy(&output.stderr)
     );
     assert!(
         !tmp.path().join("build").exists(),
